requirements.txt
comtypes==1.4.8
pynetworktables==2021.0.0
pypiwin32==223
pyttsx3==2.98
pywin32==308